blog/requirements.txt
flask>=3.0
flask-compress
markdown
pyyaml
